reorganized Side bar components to sidebar subdirectory
