+++ client/src/components/SideBar/ColorPicker.test.jsx
