+++ client/src/components/SideBar/ColorPickerContainer.test.jsx
@@ -2,10 +2,10 @@ import React from "react";
 import { expect, test, describe, beforeEach, vi } from "vitest";
 import { render, screen } from "@testing-library/react";
 import { ThemeProvider } from "@mui/material/styles";
-import theme from "../theme";
-import MinutesContext from "../contexts/MinutesContext";
+import theme from "../../theme";
+import MinutesContext from "../../contexts/MinutesContext";
 import ColorPickerContainer from "./ColorPickerContainer";
-import { mockMinutesContextState } from "../util/test.helpers";
+import { mockMinutesContextState } from "../../util/test.helpers";
 
 describe("ColorPickerContainer", () => {
   const updateMinutesMock = vi.fn();
